requirements.txt
streamlit>=1.28.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
google-generativeai>=0.3.0
requests>=2.31.0